Отрисовывает миниатюры в отдельном модуле
