+++ index.html
@@ -234,8 +234,8 @@
     </section>
   </template>
 
-  <script src="/js/draws-thumbnails.js" type="module"></script>
-  <script src="/js/main.js" type="module"></script>
+  <!-- <script src="/js/draws-thumbnails.js" type="module"></script> -->
+  <script src="js/main.js" type="module"></script>
   <!-- <script src="/js/function.js"></script> -->
 </body>
 </html>
