+++ js/main.js
@@ -1,2 +1,5 @@
 import {getKekstogramFeed} from './data.js';
-getKekstogramFeed();
+import {drowThumbnails} from './draws-thumbnails.js';
+
+const posts = getKekstogramFeed();
+drowThumbnails(posts);
